Refactor scrolling logic: add accumulator and tune sensitivity
Addresses reports of laggy and stiff scrolling behavior.

**Changes:**
- **High Precision Deltas**: Updated [MouseDelta](cci:2://file:///code/rust/stremio-linux-shell/src/shared/types.rs:18:0-18:40) in [shared/types.rs](cci:7://file:///code/rust/stremio-linux-shell/src/shared/types.rs:0:0-0:0) to use `f64` instead of `i32`, preventing the immediate truncation of small (sub-pixel) scroll events common with touchpads.
- **Scroll Accumulator**: Implemented a scroll accumulator in [WebView](cci:2://file:///code/rust/stremio-linux-shell/src/webview/mod.rs:45:0-52:1). This buffers fractional scroll values and only dispatches IPC events to the CEF engine when a whole integer threshold is reached, ensuring smooth continuous scrolling without event loss.
- **Sensitivity Tuning**:
  - Increased `LineDelta` sensitivity (mouse wheel) by **3x** (100.0 -> 300.0) to fix the "stiff" scrolling feel.
  - Increased `PixelDelta` sensitivity (touchpad) by **3x** for better responsiveness.

+++ src/shared/types.rs
@@ -16,7 +16,7 @@ pub enum Cursor {
 pub struct MousePosition(pub i32, pub i32);
 
 #[derive(Debug, Default, Clone, Copy)]
-pub struct MouseDelta(pub i32, pub i32);
+pub struct MouseDelta(pub f64, pub f64);
 
 #[derive(Debug, Clone, Copy)]
 pub struct WindowSize(pub i32, pub i32);
